requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
click==8.1.8
colorama==0.4.6
eval_type_backport==0.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
joblib==1.4.2
mistralai==1.7.0
nltk==3.9.1
numpy==1.26.4
pandas==2.2.3
pydantic==2.11.4
pydantic_core==2.33.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
scikit-learn==1.2.2
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
spicy==0.16.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
